DX12Resource remake

+++ Application/DX12ResourceUploader.h
@@ -22,7 +22,7 @@ namespace graphicsGadgetLab
 		void BeginUpload() noexcept;
 		void EndUpload(bool wait) noexcept;
 
-		void UploadResource(const void* data, size_t dataSize, const DX12Resource* destResource) noexcept;
+		void UploadResource(const void* data, uint64_t dataSize, const DX12Resource* destResource) noexcept;
 		void UploadResource(const std::vector<D3D12_SUBRESOURCE_DATA>& subResourceData, const DX12Resource* destResource) noexcept;
 
 		DX12CommandList* GetUploadCommandList() const { return m_UploadCommandList.get(); }
